requirements.txt
Pillow
accelerate
blobfile
einops
fairscale
ftfy
future
h5py
huggingface-hub
mypy
nltk
numpy
numpy==1.26.4
open_clip_torch>=0.2.1
opencv-python
pandas
pyarrow
pycocoevalcap
pytest
pytorch-lightning==2.2.1
requests
rouge
ruamel.yaml
safetensors
scikit-image
scikit-survival
scipy
sentencepiece
tensorboardX
timm==0.9.8
torch==2.0.1
torchmetrics==0.7.3
torchvision==0.15.2
transformers
wandb
webdataset>=0.2.31